% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ar_multiv.R
\name{plot_ar_multiv}
\alias{plot_ar_multiv}
\title{Function to automate the production of a set of 
time series graphs that visualise the multivariate autoregressive 
structure of two input time series}
\usage{
plot_ar_multiv(timeseriesx, timeseriesy, lags = NULL)
}
\arguments{
\item{timeseriesx}{an input time series as a vector of numerical values}

\item{timeseriesy}{an input time series as a vector of numerical values}

\item{lags}{a vector of numeric values that represent the time lags the user wants to visualise}
}
\value{
an object of class `ggplot` that represents a matrix of scatterplots
}
\description{
Function to automate the production of a set of 
time series graphs that visualise the multivariate autoregressive 
structure of two input time series
}
\author{
Trent Henderson
}
