% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ar.R
\name{plot_ar}
\alias{plot_ar}
\title{Function to automate the production of a set of 
time series graphs that visualise the autoregressive 
structure of an input time series}
\usage{
plot_ar(timeseries, lags = NULL)
}
\arguments{
\item{timeseries}{the input time series as a vector of numerical values}

\item{lags}{a vector of numeric values that represent the time lags the user wants to visualise}
}
\value{
an object of class `ggplot` that represents a matrix of scatterplots
}
\description{
Function to automate the production of a set of 
time series graphs that visualise the autoregressive 
structure of an input time series
}
\author{
Trent Henderson
}
